[Sema] add TypeCandidates<T>
