+++ src/MetroDriver/Sema/SemaDeductionFunctionType.cc
@@ -1,0 +1,14 @@
+#include "AST.h"
+#include "Error.h"
+#include "Utils.h"
+#include "MetroDriver/Sema.h"
+
+namespace metro::semantics {
+
+void Sema::create_function_dc() {
+
+}
+
+
+} // namespace metro::semantics
+
